+++ src/Samples/index.ts
@@ -111,7 +111,8 @@ let urlS2S;
             "raiseLowLevelXmppInEvent": false,
             "raiseLowLevelXmppOutReq": false,
             "maxIdleTimer": 16000,
-            "maxPingAnswerTimer": 11000
+            "maxPingAnswerTimer": 11000,
+           // "xmppRessourceName": "vnagw"
         },
         "s2s": {
             "hostCallback": urlS2S,
@@ -203,7 +204,8 @@ let urlS2S;
             // "autoLoadConversations": false,
             "autoLoadContacts": true,
             "enableCarbon": true,
-            "enablesendurgentpushmessages": true
+            "enablesendurgentpushmessages": true,
+           "useMessageEditionAndDeletionV2": true
         },
         // Services to start. This allows to start the SDK with restricted number of services, so there are less call to API.
         // Take care, severals services are linked, so disabling a service can disturb an other one.
@@ -1238,7 +1240,7 @@ let urlS2S;
             }, 10000);
     }
 
-    async function tesdeleteMessageFromConversation() {
+    async function testdeleteMessageFromConversation() {
             let that = this;
             let contactEmailToSearch = "vincent01@vbe.test.openrainbow.net";
             let contact = await rainbowSDK.contacts.getContactByLoginEmail(contactEmailToSearch);
@@ -1255,15 +1257,45 @@ let urlS2S;
                 type : "text/html"
             }; // */
             rainbowSDK.im.sendMessageToConversation(conversation, txt, "FR", content, "Le sujet de node").then(async (msgSent) => {
-                logger.log("debug", "MAIN - tesdeleteMessageFromConversation sendMessageToConversation - result : ", msgSent);
-                logger.log("debug", "MAIN - tesdeleteMessageFromConversation sendMessageToConversation - conversation : ", conversation);
+                logger.log("debug", "MAIN - testdeleteMessageFromConversation sendMessageToConversation - result : ", msgSent);
+                logger.log("debug", "MAIN - testdeleteMessageFromConversation sendMessageToConversation - conversation : ", conversation);
 
                 await Utils.until(() => {
                     return conversation.getMessageById(msgSent.id)!==undefined;
                 }, "Wait for message to be added in conversation id : " + conversation.id);
 
                 let conversationWithMessagesRemoved = await rainbowSDK.conversations.deleteMessage(conversation, msgSent.id);
-                logger.log("debug", "MAIN - tesdeleteMessageFromConversation - conversation with message removed : ", conversationWithMessagesRemoved);
+                logger.log("debug", "MAIN - testdeleteMessageFromConversation - conversation with message removed : ", conversationWithMessagesRemoved);
+
+            });
+    }
+    
+    async function testmodifyMessageFromConversation() {
+            let that = this;
+            let contactEmailToSearch = "vincent01@vbe.test.openrainbow.net";
+            let contact = await rainbowSDK.contacts.getContactByLoginEmail(contactEmailToSearch);
+            let conversation = await rainbowSDK.conversations.openConversationForContact(contact);
+            let txt = "# TYPESCRIPT in SDK for Node.JS\n" +
+                    "\n" +
+                    "Here is the howto TypeScript in **Rainbow-Node-SDK**\n";
+            let content = {
+                message: txt,
+                type: "text/markdown"
+            };
+            /*let content = {
+                message : "<a href=\"xxx\">mon lmien</<a>",
+                type : "text/html"
+            }; // */
+            rainbowSDK.im.sendMessageToConversation(conversation, txt, "FR", content, "Le sujet de node").then(async (msgSent) => {
+                logger.log("debug", "MAIN - testmodifyMessageFromConversation sendMessageToConversation - result : ", msgSent);
+                logger.log("debug", "MAIN - testmodifyMessageFromConversation sendMessageToConversation - conversation : ", conversation);
+
+                await Utils.until(() => {
+                    return conversation.getMessageById(msgSent.id)!==undefined;
+                }, "Wait for message to be added in conversation id : " + conversation.id);
+
+                let conversationWithMessagesRemoved = await rainbowSDK.conversations.deleteMessage(conversation, msgSent.id);
+                logger.log("debug", "MAIN - testmodifyMessageFromConversation - conversation with message removed : ", conversationWithMessagesRemoved);
 
             });
     }