+++ src/lib/NodeSDK.ts
@@ -44,6 +44,7 @@ let LOG_ID = "NodeSDK/IDX";
  * @property {string} options.xmpp.raiseLowLevelXmppOutReq enable the raise of event "rainbow_onxmmprequestsent" when a data is sent in xmpp pipe.
  * @property {string} options.xmpp.maxIdleTimer to define the delay without xmpp exchange after which a ping is sent to server.
  * @property {string} options.xmpp.maxPingAnswerTimer to define the time to wait the xmpp ping response.
+ * @property {string} options.xmpp.xmppRessourceName to define the name of the xmpp resource.
  * @property {string} options.s2s.hostCallback "http://3d260881.ngrok.io", S2S Callback URL used to receive events on internet.
  * @property {string} options.s2s.locallistenningport "4000", Local port where the events must be forwarded from S2S Callback Web server.
  * @property {string} options.rest.useRestAtStartup, enable the REST requests to the rainbow server at startup (used with startWSOnly method). Default value is true.
@@ -134,7 +135,8 @@ type OptionsType = {
         "raiseLowLevelXmppInEvent": false,
         "raiseLowLevelXmppOutReq": false,
         "maxIdleTimer": 15000,
-        "maxPingAnswerTimer": 10000
+        "maxPingAnswerTimer": 10000,
+        xmppRessourceName: undefined
     }
     "s2s": {
         "hostCallback": string,
@@ -338,6 +340,7 @@ class NodeSDK {
      * @param {string} options.rainbow.mode "xmpp", The event mode used to receive the events. Can be `xmpp` or `s2s` (default : `xmpp`).
      * @param {string} options.xmpp.timeBetweenXmppRequests the time between two xmpp request (avoid burst)
      * @param {string} options.xmpp.raiseLowLevelXmppInEvent enable the raise of event "rainbow_onxmmpeventreceived" when a data is received in xmpp pipe.
+     * @param {string} options.xmpp.xmppRessourceName to define the name of the xmpp resource.
      * @param {string} options.xmpp.raiseLowLevelXmppOutReq enable the raise of event "rainbow_onxmmprequestsent" when a data is sent in xmpp pipe.
      * @param {string} options.xmpp.maxIdleTimer to define the delay without xmpp exchange after which a ping is sent to server.
      * @param {string} options.xmpp.maxPingAnswerTimer to define the time to wait the xmpp ping response.